+++ engine3D/scene.h
@@ -22,8 +22,8 @@ public:
 	Scene();
 	Scene(glm::vec3 position,float angle,float hwRelation,float near, float far);
 	void addShape(int type);
-	void addShape(const std::string& fileName, glm::vec3 position, int simplifyFlag);															//@Edited
-	void addShape(const std::string& fileName,const std::string& textureFileName, int simplifyFlag);						//@Edited
+	void addShape(const std::string& fileName, glm::vec3 position, float simplifyFlag);															//@Edited
+	void addShape(const std::string& fileName,const std::string& textureFileName, float simplifyFlag);						//@Edited
 	void addShape(Vertex* vertices, unsigned int numVertices, unsigned int* indices, unsigned int numIndices);
 	void addShape(Vertex* vertices, unsigned int numVertices, unsigned int* indices, unsigned int numIndices,const std::string& textureFileName);
 	void addShader(const std::string& fileName);
